==> notebooks/01 Basic Data Cleaning.ipynb <==
import pandas as pd

# Load main database
enforce = pd.read_csv("../data/raw/enforcement.csv")

# Drop unnecessary columns
enforce = enforce.drop(['FACID', 'LTC', 'FAC_TYPE_CODE', 'PENALTY_TYPE', 'TOTAL_PENALTY_OFFSET_AMOUNT',
                        'TOTAL_COLLECTED_AMOUNT', 'TOTAL_BALANCE_DUE',
                        'ACLAIMS_VISIT_NUMBER', 'EVENTID', 'DEATH_RELATED', 'INTAKEID_ALL', 'SFY',
                        'APPEAL_DUE_DATE', 'APPEAL_RECEIVED_DATE', 'PRIORITY_ALL',
                        'PENALTY_CATEGORY_OTHER', 'VIOLATION_FROM_DATE', 'VIOLATION_TO_DATE'], axis=1)
enforce.head()

# Determine whether any columns need type conversions
for series in range(13): print(type(enforce.iloc[:, series].values[0]))

# Convert dates and currencies to numbers
enforce['PENALTY_ISSUE_DATE'] = pd.to_datetime(enforce['PENALTY_ISSUE_DATE'])
enforce['TOTAL_AMOUNT_DUE_FINAL'] = enforce['TOTAL_AMOUNT_DUE_FINAL'].replace('[\$,]', '', regex=True).astype(float)
enforce['TOTAL_AMOUNT_INITIAL'] = enforce['TOTAL_AMOUNT_INITIAL'].replace('[\$,]', '', regex=True).astype(float)

# Check for null values -- the only null values are in 'class-assessed-final', which is fine because the class doesn't
# always change before and after an appeal.

enforce.isnull().sum()

# Save cleaned database for later notebooks
enforce.to_csv("../data/interim/enforce.csv")

# Load secondary database with detailed narratives
narrative = pd.read_csv("../data/raw/citation_narr.csv", encoding='unicode-escape')
narrative.head()

# Convert facility names to uppercase to assist with merge of main database and detailed narrative database
narrative['FACILITY_NAME'] = narrative['FACILITY_NAME'].str.upper()
narrative_basics = narrative[['PENALTY_NUMBER', 'FACILITY_NAME', 'NARRATIVE']]
narrative_basics.head()

# Merge main database and narrative database

working = narrative_basics.merge(enforce, how="left", on=["PENALTY_NUMBER"])
working = working[['PENALTY_NUMBER', 'FACILITY_NAME_x', 'PENALTY_ISSUE_DATE', 'TOTAL_AMOUNT_DUE_FINAL', 'NARRATIVE']]
working.columns = ['NUM', 'FACILITY', 'DATE', 'FINE', 'NARRATIVE']
working.head()

# Check for null values in merged database
working.isnull().sum()

#See which rows are null in merged database
working[working['FINE'].isnull()]

# Drop the 2 out of 2,800 rows that are null rather than investigate why the fine is missing
working = working.drop([708, 2355])

# Save the merged database for more work in future notebooks
working.to_csv("../data/interim/working.csv")
