==> notebooks/02 Natural Language Processing.ipynb <==
import pandas as pd
import numpy as np
from sklearn.feature_extraction.text import CountVectorizer
import spacy
nlp = spacy.load('en_core_web_md')

working = pd.read_csv("../data/interim/working.csv")
working = working.drop(['Unnamed: 0'], axis=1)
working.head()

## Note: this notebook takes 30 minutes to run, primarily because of cell 5. Please feel free to skip it if you
## have downloaded the repository, as all of the work done by this notebook has been saved in the export file.

from gensim.models import Phrases

# Generate a list of strings from the database

corpus = working['NARRATIVE'].tolist()
(corpus[0][0:100], corpus[1][0:100], corpus[2][300:400],), 
(type(corpus), type(corpus[0]))

# This function will convert a string into a list of unigrams and remove whitespace, punctuation, and stopwords.

def unigrammize(input_text):
    for word in nlp.Defaults.stop_words:
        for w in (word, word[0].upper() + word[1:], word.upper()):
            lex = nlp.vocab[w]
            lex.is_stop = True
        
    lower = nlp(input_text.lower())
    unigrams = [token.lemma_ for token in lower 
                   if (not token.is_stop) and (not token.is_punct) and (not token.is_space)]
    return unigrams

%%time

# Convert the corpus into a list of lists of unigrams
unigrams = [unigrammize(document) for document in corpus]

# View a portion of the unigrams
[unigrams[0][0:10], unigrams[1][0:10], unigrams[2][0:10]]

# This function adds bigrams and trigrams to a list of list of unigrams without deleting the original unigrams.

def add_trigrams(unigrams):
    result = []
    bigram_model = Phrases(unigrams, min_count=1, delimiter=b'_')
    trigram_model = Phrases(bigram_model[unigrams], min_count=1, delimiter=b'_')
    
    for document in unigrams:
        bigrams_ = [b for b in bigram_model[document] if b.count('_') == 1]
        trigrams_ = [t for t in trigram_model[bigram_model[document]] if t.count('_') == 2]
        merged_ = trigrams_ + bigrams_ + document
        result.append(merged_)

    return result

# View a sample of the list of lists that includes trigrams

trigrams = add_trigrams(unigrams)
[trigrams[0][0:30], trigrams[1][0:30], trigrams[2][0:30], trigrams[3][0:30]]

# View another sample of the list of lists to make sure it still includes unigrams and not just trigrams
trigrams[0][40:]

# Add the trigrams to the database and save it to disk.
working.loc[:, 'TRIGRAMS'] = trigrams
working.to_csv("../data/interim/trigrams3.csv")
