==> notebooks/04 Categorical Fine Prediction.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from pylab import rcParams
rcParams['figure.figsize'] = 15, 10

from ast import literal_eval
from sklearn.feature_extraction.text import CountVectorizer

data = pd.read_csv("../data/interim/trigrams3.csv")
data = data.drop(['Unnamed: 0'], axis=1)
data['TRIGRAMS'] = data['TRIGRAMS'].apply(literal_eval)
data.head()

from sklearn.model_selection import train_test_split
from sklearn.naive_bayes import MultinomialNB
from sklearn import metrics
from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier

# Build a corpus of strings out of the trigram column in the main database

corpus = [' '.join(data.loc[row, 'TRIGRAMS']) for row in range(len(data))] 
corpus[0][0:100], corpus[1][0:100], corpus[2][0:100]

<h3>Generate the data for the $1,000 categorization problem</h3>

# Define a function that will convert the corpus into a vectorized bag of words, then apply it.
# The function also generates a binary outcome variable that is 1 if the fine was $1,000 or more.

def make_xy(data, vectorizer=None):
    #Your code here    
    if vectorizer is None:
        vectorizer = CountVectorizer()
    X = vectorizer.fit_transform(corpus)
    X = X.tocsc()  # some versions of sklearn return COO format
    y = (data.FINE > 1000).values.astype(np.int)
    return X, y
X, y = make_xy(data)

<h3 style="color:blue">Multinomial Naive Bayes predicting if fine was $1,000 or more</h3>

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.3)

foo = MultinomialNB()
foo.fit(X_train, y_train)
pred1 = foo.predict(X_train)
pred2 = foo.predict(X_test)
score1_nb1 = metrics.accuracy_score(y_train, pred1)
score2_nb1 = metrics.accuracy_score(y_test, pred2)
f1score_nb1 = metrics.f1_score(y_test, pred2)
naive_nb1 = sum(data.FINE > 1000) / sum(data.FINE > -1)

print(["training score:", score1_nb1, "test score:", score2_nb1, 
       "F1 Score", f1score_nb1, "naive:", naive_nb1])

<h3 style="color:red">Random Forest Generator predicting if fine was $1,000 or more</h3>

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.3)

foo = RandomForestClassifier(max_depth = 4, random_state=48)
foo.fit(X_train, y_train)
pred1 = foo.predict(X_train)
pred2 = foo.predict(X_test)
score1_rf1 = foo.score(X_train, y_train)
score2_rf1 = foo.score(X_test, y_test)
f1score_rf1 = metrics.f1_score(y_test, pred2)
naive_rf1 = sum(data.FINE > 1000) / sum(data.FINE > -1)

print(["training score:", score1_rf1, "test score:", score2_rf1, 
       "F1 score:", f1score_rf1, "naive:", naive_rf1])

<h3 style="color:#3cb371">Gradient Boosting Machine predicting if fine was $1,000 or more</h3>

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.3)

foo = GradientBoostingClassifier(random_state=48)
foo.fit(X_train, y_train)
pred1 = foo.predict(X_train)
pred2 = foo.predict(X_test)
score1_gb1 = foo.score(X_train, y_train)
score2_gb1 = foo.score(X_test, y_test)
f1score_gb1 = metrics.f1_score(y_test, pred2)
naive_gb1 = sum(data.FINE > 1000) / sum(data.FINE > -1)

print(["training score:", score1_gb1, "test score:", score2_gb1, "Naive Model:", naive_gb1])

<h3>Generate the data for the $5,000 categorization problem</h3>

# Define a function that will convert the corpus into a vectorized bag of words, then apply it.
# The function also generates a binary outcome variable that is 1 if the fine was $5,000 or more.

def make_xy(data, vectorizer=None):
    #Your code here    
    if vectorizer is None:
        vectorizer = CountVectorizer()
    X = vectorizer.fit_transform(corpus)
    X = X.tocsc()  # some versions of sklearn return COO format
    y = (data.FINE > 5000).values.astype(np.int)
    return X, y
X, y = make_xy(data)

<h3 style="color:blue">Multinomial Naive Bayes predicting if fine was $5,000 or more</h3>

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.3)

foo = MultinomialNB()
foo.fit(X_train, y_train)
pred1 = foo.predict(X_train)
pred2 = foo.predict(X_test)
score1_nb2 = metrics.accuracy_score(y_train, pred1)
score2_nb2 = metrics.accuracy_score(y_test, pred2)
f1score_nb2 = metrics.f1_score(y_test, pred2)
naive_nb2 = sum(data.FINE > 5000) / sum(data.FINE > -1)

print(["training score:", score1_nb2, "test score:", score2_nb2, 
       "F1 Score", f1score_nb2, "naive:", naive_nb2])

<h3 style="color:red">Random Forest Generator predicting if fine was $5,000 or more</h3>

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.3)

foo = RandomForestClassifier(max_depth = 4, random_state=48)
foo.fit(X_train, y_train)
pred1 = foo.predict(X_train)
pred2 = foo.predict(X_test)
score1_rf2 = foo.score(X_train, y_train)
score2_rf2 = foo.score(X_test, y_test)
f1score_rf2 = metrics.f1_score(y_test, pred2)
naive_rf2 = sum(data.FINE > 5000) / sum(data.FINE > -1)

print(["training score:", score1_rf2, "test score:", score2_rf2,
       "F1 score:", f1score_rf2, "naive:", naive_rf2])

<h3 style="color:#3cb371">Gradient Boosting Machine predicting if fine was $5,000 or more</h3>

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.3)

foo = GradientBoostingClassifier(random_state=48)
foo.fit(X_train, y_train)
pred1 = foo.predict(X_train)
pred2 = foo.predict(X_test)
score1_gb2 = foo.score(X_train, y_train)
score2_gb2 = foo.score(X_test, y_test)
f1score_gb2 = metrics.f1_score(y_test, pred2)
naive_gb2 = sum(data.FINE > 5000) / sum(data.FINE > -1)

print(["training score:", score1_gb2, "test score:", score2_gb2, 
       "F1 Score", f1score_nb2, "Naive Model:", naive_gb2])

<h3>Summary of performance of categorical models</h3>

summary = pd.DataFrame(
    {"MultinomialNB": [score1_nb1, score2_nb1, f1score_nb1, naive_nb1,
                      score1_nb2, score2_nb2, f1score_nb2, naive_nb2],
    "RandomForest": [score1_rf1, score2_rf1, f1score_rf1, naive_rf1,
                    score1_rf2, score2_rf2, f1score_rf2, naive_rf2],
    "GradientBoost": [score1_gb1, score2_gb1, f1score_gb1, naive_gb1,
                     score1_gb2, score2_gb2, f1score_gb2, naive_gb2]})
summary.index = ['Train >$1,000', 'Test >$1,000', 'F1 >$1,000', 'Naive >$1,000',
                'Train >$5,000', 'Test >$5,000', 'F1 >$5,000', 'Naive >$5,000']
summary.round(3)
