==> notebooks/06 Topic Modeling.ipynb <==
import pandas as pd
import numpy as np
import spacy

from gensim.models import Phrases, LdaModel, CoherenceModel, Word2Vec
from gensim.models.word2vec import LineSentence
from gensim.corpora import Dictionary, MmCorpus

from ast import literal_eval

data = pd.read_csv("../data/interim/trigrams3.csv")
data['TRIGRAMS'] = data['TRIGRAMS'].apply(literal_eval)
data.head()

# Generate a sample of 25% of the database

sample = data.sample(n=700, random_state=42)
#sample = [trigram_list for trigram_list in sample['TRIGRAMS']]
sample['TRIGRAMS']

slist = sample['TRIGRAMS'].tolist()
print(type(slist), type(slist[0]), type(slist[0][0]))

%%time
# Create a dictionary with all of the words in the sample corpus
corpus = slist
dct = Dictionary(corpus)
dct.filter_extremes(no_below=2, no_above=0.15)

%%time
# Convert the dictionary to numeric wordIDs
corpus_nums = [dct.doc2bow(text) for text in corpus]

%%time
# Build a model using the numeric wordIDs, letting the function know how to convert these id's back into words
lda_model = LdaModel(corpus_nums, num_topics=10, id2word=dct, passes=30, iterations=600, random_state=42)

# Print the most commonly occurring words in each topic

for topic in range(10):
    print(lda_model.show_topic(topic, topn=15))
    print("\n")


# Define an informal grid of hyperparameters to test

topicnums = [3, 5, 8, 10, 20]
alphas = [0.01, 0.1, 0.3, 0.9]
etas = [0.01, 0.1, 0.3, 0.9]

# Evaluate each set of hyperparameters based on its c_v coherence score; this is time-consuming and has been set to 'off'
# by default.

hyperTest = False

if hyperTest:
    results = {}
    for t in topicnums:
        for a in alphas:
            for e in etas:
                lda_model = LdaModel(corpus_nums, id2word=dct, passes=30, iterations=600, random_state=42,
                                     num_topics=t, alpha=a, eta=e)
                coherence_model_lda = CoherenceModel(model=lda_model, texts=corpus, dictionary=dct, coherence='c_v')
                results[(t, a, e)] = coherence_model_lda.get_coherence()
    import csv
    w = csv.writer(open("output.csv", "w"))
    for key, val in results.items():
        w.writerow([key, val])

printout = pd.read_csv("output.csv", names=['Topics, Alpha, Eta', 'Coherence Score'])
printout.sort_values('Coherence Score', ascending=False).head(20)
