==> notebooks/07 Topic Visualization.ipynb <==
import pandas as pd
import numpy as np
import spacy
import re
import pickle

from gensim.models import Phrases, LdaModel, CoherenceModel, Word2Vec
from gensim.models.word2vec import LineSentence
from gensim.corpora import Dictionary, MmCorpus
from gensim.test.utils import datapath

import pyLDAvis
import pyLDAvis.gensim

from ast import literal_eval

data = pd.read_csv("../data/interim/trigrams3.csv")
data['TRIGRAMS'] = data['TRIGRAMS'].apply(literal_eval)
data.head()

# Generate a sample of 80% of the database (holding out 20% for validation)

train = data.sample(frac=0.8, random_state=42)
test = data.drop(train.index)
train['TRIGRAMS']

test['TRIGRAMS']

def remove_numbers(corpus):
    regexp = re.compile(r'^[0-9]*$')
    return [[word for word in line if not regexp.search(word)] for line in corpus]

# Fill a dictionary with all of the words and phrases in the corpus, except for pure numbers like "2016", which
# do not contain much valuable info about what topic a narrative belongs to. Filter out words that occur too often
# or not often enough. Then convert the dictionary to vectors.

corpus = train['TRIGRAMS'].tolist()
corpus = remove_numbers(corpus)
dct = Dictionary(corpus)
dct.filter_extremes(no_below=2, no_above=0.15)
corpus_nums = [dct.doc2bow(text) for text in corpus]

# Train an LDA model, somewhat more carefully, using the hyperparameters that were close to optimal in the previous
# notebook. Although a 3-topic model had slightly more coherence, that is less interesting than an 8-topic model,
# and the coherence statistics were very similar.

retrainModel = False

if retrainModel:
    lda_model = LdaModel(corpus_nums, num_topics=8, id2word=dct, passes=60, iterations=1200, 
                         alpha=0.9, eta=0.9, random_state=42)
    lda_model.save("../models/ldatrain")
    dct.save_as_text("../models/ldatrain_dct")

lda_model = LdaModel.load("../models/ldatrain")
dct = Dictionary.load_from_text("../models/ldatrain_dct")

# Preview the most-frequently occuring words in each of the 8 topics, showing the percentage of each topic represented
# by each word, e.g., "oxygen, 0.51" means oxygen is 0.51% of the words in the topic.

for topic in range(8):
    print([(a, round(b*100, 2)) for (a, b) in lda_model.show_topic(topic, topn=15)])
    print("\n")

%%time

# Prepare a graphical summary of the topics based on principal component analysis (very time consuming)
PrepVisual = False

if PrepVisual:
    LDAvis_prepared = pyLDAvis.gensim.prepare(lda_model, corpus_nums, dct)
    pickle.dump( LDAvis_prepared, open( "../models/visual.p", "wb" ) )
    pyLDAvis.save_html(LDAvis_prepared, '../models/lda.html')




LDAvis_prepared = pickle.load( open( "../models/visual.p", "rb" ) )


# Launch an interactive visualization of the topics in a new window (using this window distorts the width of all cells)
pyLDAvis.show(LDAvis_prepared)
