==> notebooks/08 Topic Prediction with Train and Test Data.ipynb <==
import pandas as pd
import numpy as np
import spacy
import re
from scipy import stats

from gensim.models import Phrases, LdaModel, CoherenceModel, Word2Vec
from gensim.models.word2vec import LineSentence
from gensim.corpora import Dictionary, MmCorpus
from gensim.test.utils import datapath

import pyLDAvis
import pyLDAvis.gensim

from ast import literal_eval

# Import model from disc
lda_model =  LdaModel.load("../models/ldatrain")
dct = Dictionary.load_from_text("../models/ldatrain_dct")

# Test that import was successful
for topic in range(2):
    print([(a, round(b*100, 2)) for (a, b) in lda_model.show_topic(topic, topn=5)])
    print("\n")

data = pd.read_csv("../data/interim/trigrams3.csv")
data = data.drop(['Unnamed: 0'], axis=1)
data['TRIGRAMS'] = data['TRIGRAMS'].apply(literal_eval)
data.head()

def remove_numbers(corpus):
    regexp = re.compile(r'^[0-9]*$')
    return [[word for word in line if not regexp.search(word)] for line in corpus]
full_corpus = data['TRIGRAMS'].tolist()
full_corpus = remove_numbers(full_corpus)
full_dct = dct
full_corpus_nums = [full_dct.doc2bow(text) for text in full_corpus]

# Test that topics have actually been assigned to documents
lda_model.get_document_topics(full_corpus_nums[2])

# Write a function to identify the topic most strongly associated with each document

def get_best_topic(document, model):
    best_topic = None
    best_prob = 0
    candidates = model.get_document_topics(document)
    for (topic, prob) in candidates:
        if prob > best_prob:
            best_topic = topic
            best_prob = prob
    return best_topic

# Test the function
get_best_topic(full_corpus_nums[2], lda_model)

get_best_topic(full_corpus_nums[1828], lda_model)

%%time

# Assign each document in the corpus to a particular topic
data.loc[:, 'TOPIC'] = [get_best_topic(full_corpus_nums[row], lda_model) 
                        for row in range(len(data))]

data.head()

train = data.sample(frac=0.8, random_state=42)
test = data.drop(train.index)
train.head()

test.head()

train_means = train.groupby(['TOPIC']).mean()
train_means

test_means = test.groupby(['TOPIC']).mean()
test_means

compare_means = pd.DataFrame({"Train": train_means['FINE'], "Test": test_means['FINE']})
compare_means

# Display average fines by topic for training set vs. test set
compare_means.index=['Life Support', 'Escape', 'Sores', 'Administration', 'Abuse', 'Theft', 'Diabetes', 'Rashes']
pd.options.display.float_format = '${:,.0f}'.format
compare_means.sort_values('Train', ascending=False)


# Average fines are very tightly correlated bewteen topics, r=0.96
stats.pearsonr(compare_means['Train'], compare_means['Test'])
